HW5.Action

Documentation

newtype HIO a

IO for the Hi language takes in a set of the users permissions and returns an IO operation

Constructors

HIO

Fields

runHIO :: Set HiPermission -> IO a

data HiPermission

Permissions for readwritetime operations

Constructors

AllowRead

AllowWrite

AllowTime

newtype PermissionException

Constructors

PermissionRequired HiPermission