HW5.Evaluator

Documentation

eval :: HiMonad m => HiExpr -> m (Either HiError HiValue)

Basic Logic of Evaluation of Expressions of the Hi Language. - Implementation details are abstracted away to Impl.hs

The main point of entry for our evaluator, it pattern matches on the value of HiExpr and depending on if the expression is a value, action, index application (on strings, lists, bytestrings) returns the corresponding value equating to it.